+++ lib/regexpattern.dart
@@ -1,2 +1,3 @@
 export 'package:regexpattern/src/regex_pattern.dart';
-export 'package:regexpattern/src/regex_validation.dart';
+export 'package:regexpattern/src/regex_validator.dart';
+export 'package:regexpattern/src/regex_extension.dart';