+++ lib/src/regex_validator.dart
@@ -1,4 +1,5 @@
-class RegexValidation {
+/// Regex Validator
+class RegVal {
   /// Returns whether the pattern has a match in the string [input].
   static bool hasMatch(String s, Pattern p) =>
       (s == null) ? false : RegExp(p).hasMatch(s);